pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentux"
version = "0.1.0"
description = "Synthetic observability for agent usability"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [{ name = "AgentUX Contributors" }]
keywords = ["agent", "ux", "observability", "benchmark", "ai", "synthetic-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
    "Environment :: Console",
]
dependencies = [
    "textual>=0.79.0",
    "rich>=13.9.0",
    "typer[all]>=0.12.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "playwright>=1.48.0",
    "httpx>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0.0",
    "markdown-it-py>=3.0.0",
    "openai>=1.50.0",
    "anthropic>=0.37.0",
    "apscheduler>=3.10.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]

[project.scripts]
agentux = "agentux.cli.app:main"

[project.urls]
Homepage = "https://github.com/Rudrajmehta/AgentUX"
Repository = "https://github.com/Rudrajmehta/AgentUX"
Documentation = "https://github.com/Rudrajmehta/AgentUX/tree/main/docs"
Issues = "https://github.com/Rudrajmehta/AgentUX/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/agentux"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["agentux"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
